+++ src/include/index/vamana_group.h
@@ -142,13 +142,13 @@ class vamana_index_group : public base_index_group<index_type> {
   uint64_t get_l_build() const {
     return metadata_.l_build_;
   }
-  void set_l_build(uint64_t l_build) {
+  void set_l_build(uint32_t l_build) {
     metadata_.l_build_ = l_build;
   }
   uint64_t get_r_max_degree() const {
     return metadata_.r_max_degree_;
   }
-  void set_r_max_degree(uint64_t r_max_degree) {
+  void set_r_max_degree(uint32_t r_max_degree) {
     metadata_.r_max_degree_ = r_max_degree;
   }
   float get_alpha_min() const {
