+++ src/include/test/unit_flat_pq_index.cc
@@ -821,7 +821,7 @@ TEST_CASE("flat_pq_index write and read", "[flat_pq_index]") {
   const bool debug = false;
 
   uint64_t dimensions_{128};
-  uint64_t num_subspaces_{16};
+  uint32_t num_subspaces_{16};
   uint32_t bits_per_subspace_{8};
   uint32_t num_clusters_{256};
